[OPP-1405] Flytter endepunkter fra gammel personcontroller og sletter gammel personcontroller
Sletter også tester til gammel personcontroller.
